apps/sitefit/worker-fastapi/requirements.txt
# Worker dependencies

# Azure Service Bus
azure-servicebus>=7.11.0

# Database
psycopg2-binary>=2.9.9

# HTTP client (for calling AppServer)
httpx>=0.26.0

# Async support
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
